Split implementation into multiple files
