+++ include/snitch/snitch_error_handling.hpp
@@ -7,6 +7,9 @@
 #include <string_view>
 
 namespace snitch {
+// Maximum length of error messages.
+constexpr std::size_t max_message_length = SNITCH_MAX_MESSAGE_LENGTH;
+
 [[noreturn]] void terminate_with(std::string_view msg) noexcept;
 
 extern small_function<void(std::string_view)> assertion_failed_handler;
