+++ src/snitch_console.cpp
@@ -1,0 +1,9 @@
+#include "snitch/snitch_console.hpp"
+
+#include <cstdio> // for std::fwrite
+
+namespace snitch::impl {
+void stdout_print(std::string_view message) noexcept {
+    std::fwrite(message.data(), sizeof(char), message.length(), stdout);
+}
+} // namespace snitch::impl
