+++ src/snitch_matcher.cpp
@@ -1,0 +1,22 @@
+#include "snitch/snitch_matcher.hpp"
+
+namespace snitch::matchers {
+contains_substring::contains_substring(std::string_view pattern) noexcept :
+    substring_pattern(pattern) {}
+
+bool contains_substring::match(std::string_view message) const noexcept {
+    return message.find(substring_pattern) != message.npos;
+}
+
+small_string<max_message_length>
+contains_substring::describe_match(std::string_view message, match_status status) const noexcept {
+    small_string<max_message_length> description_buffer;
+    append_or_truncate(
+        description_buffer, (status == match_status::matched ? "found" : "could not find"), " '",
+        substring_pattern, "' in '", message, "'");
+    return description_buffer;
+}
+
+with_what_contains::with_what_contains(std::string_view pattern) noexcept :
+    contains_substring(pattern) {}
+} // namespace snitch::matchers
